fix for a possible memory corruption
git-svn-id: http://svn.miranda-ng.org/main/trunk@12458 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/NewXstatusNotify/src/indsnd.cpp
@@ -63,7 +63,7 @@ TCHAR *SelectSound(HWND hwndDlg, TCHAR *buff, size_t bufflen)
 	OPENFILENAME ofn = { 0 };
 
 	HWND hList = GetDlgItem(hwndDlg, IDC_INDSNDLIST);
-	ListView_GetItemText(hList, ListView_GetNextItem(hList, -1, LVNI_SELECTED), 1, buff, bufflen);
+	ListView_GetItemText(hList, ListView_GetNextItem(hList, -1, LVNI_SELECTED), 1, buff, (DWORD)bufflen);
 	if (!mir_tstrcmp(buff, TranslateT(DEFAULT_SOUND)))
 		buff = NULL;
